Port `TryFromSpec` to new golden test setup
