+++ plutarch-testlib/goldens/data-verif.uplc.golden
